requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
pyyaml>=6.0
duckdb>=0.9.0
matplotlib>=3.7.0
mplsoccer>=1.1.0
